This is a major change to the layout and styling of the app.
Reworked the layout to proper use of bootstrap cards.

Put the data filter controls on the bottom in their own card.

Added a modal dialogue box that appears on page load. But,
can be clicked away. This allows a much greater focus on the
application.

Modified the margins on the snow SWE plot so there is less white space.

Set the defaults of the radio buttons to require 20+ years and active
stations by default.

Signed-off-by: Faron Anslow <faron.anslow@nrcan-rncan.gc.ca>